apply corrections
https://coveord.atlassian.net/browse/CDX-252
