+++ packages/cli-e2e/package.json
@@ -42,6 +42,7 @@
     "async-retry": "1.3.1",
     "axios": "^0.21.1",
     "babel-jest": "26.6.3",
+    "devtools-protocol": "0.0.878340",
     "fs-extra": "^9.1.0",
     "jest": "26.6.3",
     "puppeteer": "^8.0.0",
